+++ src/pages/søknad/kvittering/kvittering-nb.ts
@@ -1,11 +1,13 @@
 export default {
+    'feil.feilOppsto': 'En feil oppsto',
+
     'kvittering.søknadMottatt': 'Søknaden om supplerende stønad er mottatt.',
     'kvittering.skrivUtSøknad': 'Skriv ut søknad',
     'kvittering.avslutt': 'Avslutt',
     'knapp.tilbake': 'Tilbake',
 
     'kvittering.saksnummer': 'Saksnummer: {saksnummer}',
-
+    'kvittering.søknadSendtInn': 'Søknad for {navn} er nå sendt inn',
     'kvittering.tilVeileder.heading': 'Til veileder',
     'kvittering.tilVeileder.punkt1': 'Skriv ut to eksempler av søknaden',
     'kvittering.tilVeileder.punkt2': 'Få underskrift av søker på den ene kopien og gi den andre kopien til søker',
