--- src/pages/søknad/steg/kvittering/kvittering.module.less
@@ -1,39 +1,0 @@
-@import '~/styles/variables.less';
-
-.suksessContainer {
-    display: flex;
-    justify-content: center;
-
-    > :not(:last-child) {
-        margin-right: @spacing-xxl;
-    }
-}
-
-.nySøknadKnapp {
-    display: flex;
-    justify-content: center;
-    flex-direction: row;
-    margin-top: @spacing-xl;
-
-    > :not(:last-child) {
-        margin-right: @spacing;
-    }
-}
-
-.senderSøknadSpinnerContainer {
-    display: flex;
-    justify-content: center;
-}
-
-.tilVeileder {
-    margin-top: @spacing-l;
-    h3 {
-        margin-bottom: @spacing-xs;
-    }
-    ol {
-        list-style: decimal inside;
-        li {
-            margin-bottom: @spacing-xs;
-        }
-    }
-}
